machine_learning/requirements.txt
torch>=2.1.0
torchvision>=0.18.0
transformers>=4.40.0
timm>=0.9.0
sentencepiece>=0.1.99
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.26.0
